+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- Changed: base theme now provides a less strict eslint config
+  (only eslint + eslint-plugin-react recommended rules)
+
 # 0.17.8 - 2016-10-08
 
 - Fixed: Service worker file was imported relatively, which only works at the